[FIX] Fix answer when no or only one object detected

+++ ayesaac/services/text_to_speech/main.py
@@ -17,9 +17,10 @@ class TextToSpeech(object):
 
     def callback(self, body, **_):
         pprint(body)
-        gTTS(text=body['response'], lang='en', slow=False).save("audio.mp3")
-        playsound('audio.mp3')
-        os.remove('audio.mp3')
+        if body['response']:
+            gTTS(text=body['response'], lang='en', slow=False).save("audio.mp3")
+            playsound('audio.mp3')
+            os.remove('audio.mp3')
         body['path_done'].append(self.__class__.__name__)
 
     def run(self):
